fast_api/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
build==1.2.2.post1
certifi==2025.4.26
charset-normalizer==3.4.3
click==8.1.8
contourpy==1.3.3
cycler==0.12.1
fastapi==0.115.12
fonttools==4.59.2
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
joblib==1.4.2
kiwisolver==1.4.9
matplotlib==3.10.6
numpy==2.2.4
packaging==24.2
pandas==2.3.2
pillow==11.3.0
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
seaborn==0.13.2
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.0
wheel==0.45.1
